svc/timesync: Convert to 96MHz PMIC REFCLK_MAIN PLL
For the purposes of TimeSync we really need the AP and SVC timer clocks to
tick at the same rate. To achieve this - it is necessary to enable the PMIC
supplied REFCLK_MAIN as the input to the PLL via the HSE input. Once done
it is possible to setup peripheral timers on the APB2 timer clocks at an
even division of the 96MHz reference clock to get to a 19.2MHz fundamental
tick in the timer. Using the same clock rate based on the same input means
that clocks don't drift when doing integer multiplication/division
conversion of clocks to nanoseconds.

Testing Done: Compiled, tested, booted.
Bug Number: TBD
Review URL:
Reviewed by: TBD

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>